move some more things
