+++ src/twisted/mail/pop3.py
@@ -16,9 +16,10 @@ import binascii
 import warnings
 from hashlib import md5
 
-from zope.interface import implementer, Interface
+from zope.interface import implementer
 
 from twisted.mail import smtp
+from twisted.mail.interfaces import IServerFactory, IMailbox
 from twisted.protocols import basic
 from twisted.protocols import policies
 from twisted.internet import task
@@ -1331,168 +1332,6 @@ class POP3(basic.LineOnlyReceiver, policies.TimeoutMixin):
                 IMailbox
             )
         raise cred.error.UnauthorizedLogin()
-
-
-
-class IServerFactory(Interface):
-    """
-    An interface for querying capabilities of a POP3 server.
-
-    Any cap_* method may raise L{NotImplementedError} if the particular
-    capability is not supported.  If L{cap_EXPIRE()} does not raise
-    L{NotImplementedError}, L{perUserExpiration()} must be implemented,
-    otherwise they are optional.  If L{cap_LOGIN_DELAY()} is implemented,
-    L{perUserLoginDelay()} must be implemented, otherwise they are optional.
-
-    @type challengers: L{dict} of L{bytes} -> L{IUsernameHashedPassword
-        <cred.credentials.IUsernameHashedPassword>}
-    @ivar challengers: A mapping of challenger names to
-        L{IUsernameHashedPassword <cred.credentials.IUsernameHashedPassword>}
-        provider.
-    """
-    def cap_IMPLEMENTATION():
-        """
-        Return a string describing the POP3 server implementation.
-
-        @rtype: L{bytes}
-        @return: Server implementation information.
-        """
-
-
-    def cap_EXPIRE():
-        """
-        Return the minimum number of days messages are retained.
-
-        @rtype: L{int} or L{None}
-        @return: The minimum number of days messages are retained or none, if
-            the server never deletes messages.
-        """
-
-
-    def perUserExpiration():
-        """
-        Indicate whether the message expiration policy differs per user.
-
-        @rtype: L{bool}
-        @return: C{True} when the message expiration policy differs per user,
-            C{False} otherwise.
-        """
-
-
-    def cap_LOGIN_DELAY():
-        """
-        Return the minimum number of seconds between client logins.
-
-        @rtype: L{int}
-        @return: The minimum number of seconds between client logins.
-        """
-
-
-    def perUserLoginDelay():
-        """
-        Indicate whether the login delay period differs per user.
-
-        @rtype: L{bool}
-        @return: C{True} when the login delay differs per user, C{False}
-            otherwise.
-        """
-
-
-
-class IMailbox(Interface):
-    """
-    An interface for mailbox access.
-
-    Message indices are 0-based.
-
-    @type loginDelay: L{int}
-    @ivar loginDelay: The number of seconds between allowed logins for the
-        user associated with this mailbox.
-
-    @type messageExpiration: L{int}
-    @ivar messageExpiration: The number of days messages in this mailbox will
-        remain on the server before being deleted.
-    """
-    def listMessages(index=None):
-        """
-        Retrieve the size of a message, or, if none is specified, the size of
-        each message in the mailbox.
-
-        @type index: L{int} or L{None}
-        @param index: The 0-based index of the message.
-
-        @rtype: L{int}, sequence of L{int}, or L{Deferred <defer.Deferred>}
-        @return: The number of octets in the specified message, or, if an
-            index is not specified, a sequence of the number of octets for
-            all messages in the mailbox or a deferred which fires with
-            one of those. Any value which corresponds to a deleted message
-            is set to 0.
-
-        @raise ValueError or IndexError: When the index does not correspond to
-            a message in the mailbox.  The use of ValueError is preferred.
-        """
-
-
-    def getMessage(index):
-        """
-        Retrieve a file containing the contents of a message.
-
-        @type index: L{int}
-        @param index: The 0-based index of a message.
-
-        @rtype: file-like object
-        @return: A file containing the message.
-
-        @raise ValueError or IndexError: When the index does not correspond to
-            a message in the mailbox.  The use of ValueError is preferred.
-        """
-
-
-    def getUidl(index):
-        """
-        Get a unique identifier for a message.
-
-        @type index: L{int}
-        @param index: The 0-based index of a message.
-
-        @rtype: L{bytes}
-        @return: A string of printable characters uniquely identifying the
-            message for all time.
-
-        @raise ValueError or IndexError: When the index does not correspond to
-            a message in the mailbox.  The use of ValueError is preferred.
-        """
-
-
-    def deleteMessage(index):
-        """
-        Mark a message for deletion.
-
-        This must not change the number of messages in this mailbox.  Further
-        requests for the size of the deleted message should return 0.  Further
-        requests for the message itself may raise an exception.
-
-        @type index: L{int}
-        @param index: The 0-based index of a message.
-
-        @raise ValueError or IndexError: When the index does not correspond to
-            a message in the mailbox.  The use of ValueError is preferred.
-        """
-
-
-    def undeleteMessages():
-        """
-        Undelete all messages marked for deletion.
-
-        Any message which can be undeleted should be returned to its original
-        position in the message sequence and retain its original UID.
-        """
-
-
-    def sync():
-        """
-        Discard the contents of any message marked for deletion.
-        """
